Create rock paper scissors
playerPlay() requests user input and converts to lowercase.
Then compares if that input is valid. If valid assigns playerInput
the inputCheck value (lowercase input).

computerPlay() randomly chooses from a list and assigns
compInput the value.

compare() takes playInput and compInput. Then proceeds to check
if the result is a tie, loss, or win. Depending on result the
score for the winning player is incremented by 1.

game() loops through game 5 times and then announces winner of game.